refactor: move controlplane to root package
