+++ onr-admin/internal/cli/access_key.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/r9s-ai/open-next-router/onr-admin/internal/store"
-	"github.com/r9s-ai/open-next-router/onr-core/pkg/controlplane"
 	"github.com/r9s-ai/open-next-router/onr-core/pkg/keystore"
 	"github.com/r9s-ai/open-next-router/pkg/config"
+	"github.com/r9s-ai/open-next-router/pkg/controlplane"
 	"github.com/spf13/cobra"
 )
