+++ onr-admin/internal/tui/service.go
@@ -8,9 +8,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/r9s-ai/open-next-router/onr-core/pkg/controlplane"
 	"github.com/r9s-ai/open-next-router/onr-core/pkg/keystore"
 	"github.com/r9s-ai/open-next-router/pkg/config"
+	"github.com/r9s-ai/open-next-router/pkg/controlplane"
 )
 
 type adminService struct {
