+++ onr/internal/onrserver/state.go
@@ -3,9 +3,9 @@ package onrserver
 import (
 	"sync"
 
-	"github.com/r9s-ai/open-next-router/onr-core/pkg/controlplane"
 	"github.com/r9s-ai/open-next-router/onr-core/pkg/keystore"
 	"github.com/r9s-ai/open-next-router/onr-core/pkg/models"
+	"github.com/r9s-ai/open-next-router/pkg/controlplane"
 )
 
 type state struct {
